third_party/WebKit/LayoutTests/fast/dom/document-set-body.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../js/resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Tests setting document.body");

iframe1 = document.createElement('iframe');
document.body.appendChild(iframe1);
document1 = iframe1.contentDocument.implementation.createHTMLDocument("document");

shouldThrow("document1.body = iframe1", "'HierarchyRequestError: A Node was inserted somewhere it doesn\\'t belong.'");
shouldBe("iframe1.parentNode", "document.body");

shouldThrow("document1.body = document1.createElement('iframe')", "'HierarchyRequestError: A Node was inserted somewhere it doesn\\'t belong.'");

document1.body = document.body;
shouldBeTrue("document1.body != document.body");

body1 = document1.createElement('body');
document1.body = body1;
shouldBe("document1.body", "body1")
</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
