third_party/WebKit/LayoutTests/fast/dom/document-set-body-expected.txt

Tests setting document.body

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document1.body = iframe1 threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS iframe1.parentNode is document.body
PASS document1.body = document1.createElement('iframe') threw exception HierarchyRequestError: A Node was inserted somewhere it doesn't belong..
PASS document1.body != document.body is true
PASS document1.body is body1
PASS successfullyParsed is true

TEST COMPLETE
